// src/themes/types.ts
export type ThemeOptionsVars = {
  fontFamily: string
  borderRadius: string
  colors: CSSModuleClasses
  heading: string
  paper: string
  backgroundDefault: string
  background: string
  darkTextPrimary: string
  darkTextSecondary: string
  textDark: string
  menuSelected: string
  menuSelectedBack: string
  divider: string
}